tests: Output comparison is part of the TestRunner
